Harden Clojars user management security
== Set password reset code by user, not username/email

Currently, email addresses are not validated and users don't need to
confirm their email before logging in. This means an attacker can set
their email to be another users username. Resetting a password
accepts a username OR email, and creates reset tokens for any user or
email that matches.

An the attacker could reset their password by providing their 'email'
which is another users username, and a password reset token
would be generated for both their user, and the target. Only one email
will be sent though, and it will either go to the target, or the invalid
email address. Clojars will throw an exception when trying to send an
email to an address that isn't valid, so there is no way for the
attacker to get the reset code.

However, if they did get it somehow, or the user reset their password
from that link, then the passwords for BOTH accounts would be reset, as
Clojars updates the password by reset code, without scoping it to the
intended user. AFAICT this isn't exploitable, but it feels close.

This patch
- Looks up the user by username or email, and then sets the reset code
  based on the user found. This prevents the reset code being assigned
to two users.
- When resetting the password, looks up the user by reset code, with
  LIMIT 1, and then adds an additional WHERE clause to only set that
  users password.
- Renames symbols and keywords named update-password or variants thereof
  to be called reset-password, to make the intent of this code clearer
  and avoid accidental coding mistakes in the future.

== Validate email addresses, ensure new passwords are 8 characters or
   more

The previous issue showed that an attacker could set their email to be
the same as a users username.

This patch
- Enforces that email addresses must have an @ sign with characters on
  both sides of the @ sign. Usernames are validated to not have an @
  sign, so this collision shouldn't be able to happen anymore.
- Sets an 8 character minimum password length, for greater security if
  user passwords are compromised.
- Refactors how validations are combined to ensure that they are not
  duplicated, and each of the following places accepting user updates
applies the right validations:
  - New user
  - Updating user details but not password
  - Updating user details including password
  - Resetting password

== Trim trailing spaces from email address

This is just good hygiene to trim whitespace from an email address.

== Only allow login through username, not email address

This one is debatable, given the changes that have been made above.
However, the find-user-by-username-or-email has caused unexpected issues
before, and it might be safer to only allow people to login through
their username.